rag/pyproject.toml
[project]
name = "rag-core"
version = "0.1.0"
description = "Mini RAG pipeline with hybrid BM25 + dense retrieval"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pydantic",
    "python-json-logger",
    "sentence-transformers",
    "numpy",
    "openai",
    "rank-bm25",
    "langchain-text-splitters",
    "pymupdf",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_core"]
